playground/requirements.txt
google-cloud-language>=1.3.0
beautifulsoup4>=4.9.1
gensim>=3.8.3
urllib3>=1.25.10
requests>=2.24.0
